--- libs/platform/src/lib/components/form/input-group/addon/index.ts
@@ -1,3 +1,0 @@
-export { InputGroupAddonBodyComponent } from './addon-body.component';
-export { InputGroupAddonComponent } from './addon.component';
-export { InputGroupInputComponent } from './input.component';
