+++ libs/platform/src/lib/components/form/input-group/input-group.module.ts
@@ -6,7 +6,9 @@ import { PlatformButtonModule } from '../../button/button.module';
 import { PlatformInputModule } from '../input/fdp-input.module';
 
 import { InputGroupComponent } from './input-group.component';
-import { InputGroupAddonComponent, InputGroupAddonBodyComponent, InputGroupInputComponent } from './addon';
+import { InputGroupAddonComponent } from './addon.component';
+import { InputGroupAddonBodyComponent } from './addon-body.component';
+import { InputGroupInputComponent } from './input.component';
 
 @NgModule({
     declarations: [
