+++ libs/platform/src/lib/components/form/input-group/public_api.ts
@@ -1,4 +1,5 @@
 export { PlatformInputGroupModule } from './input-group.module';
 export { InputGroupConfig } from './input-group.config';
 export { InputGroupComponent } from './input-group.component';
-export { InputGroupAddonComponent, InputGroupInputComponent } from './addon';
+export { InputGroupInputComponent } from './input.component';
+export { InputGroupAddonComponent } from './addon.component';
